+++ src/testFixtures/java/org/opensearch/knn/KNNResult.java
@@ -5,20 +5,41 @@
 
 package org.opensearch.knn;
 
-public class KNNResult {
-    private String docId;
-    private Float[] vector;
+import java.util.Arrays;
+import java.util.Objects;
 
-    public KNNResult(String docId, Float[] vector) {
+public class KNNResult {
+    private final static float delta = 1e-3f;
+
+    private String docId;
+    private float[] vector;
+    private Float score;
+
+    public KNNResult(String docId, float[] vector, Float score) {
         this.docId = docId;
         this.vector = vector;
+        this.score = score;
     }
 
     public String getDocId() {
         return docId;
     }
 
-    public Float[] getVector() {
+    public float[] getVector() {
         return vector;
+    }
+
+    public Float getScore() {
+        return score;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        KNNResult knnResult = (KNNResult) o;
+        return Objects.equals(docId, knnResult.docId)
+            && Arrays.equals(vector, knnResult.vector)
+            && (Float.compare(score, knnResult.score) == 0 || Math.abs(score - knnResult.score) <= delta);
     }
 }