+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Features
 ### Enhancements
 * Make the HitQueue size more appropriate for exact search [#1549](https://github.com/opensearch-project/k-NN/pull/1549)
+* Support script score when doc value is disabled [#1573](https://github.com/opensearch-project/k-NN/pull/1573)
 ### Bug Fixes
 ### Infrastructure
 ### Documentation